+++ libs/rust/sdk/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Optional `ActionHook`s to provide store-specific implementation of SDK actions.
+- Ping action (`agent.replicante.io/test.ping`).
 
 ### Changed
 - **BREAKING**: Removed Agent method to expose the graceful stop action.